chore: separted utils and command utils for future split of the libraries.
