+++ src/cli/commands/mc-agent-token.ts
@@ -4,17 +4,8 @@ import * as fs from "fs";
 import * as jwt from "jsonwebtoken";
 import * as path from "path";
 import { MindConnectAgent } from "../..";
-import {
-    checkCertificate,
-    errorLog,
-    getHomeDotMcDir,
-    homeDirLog,
-    proxyLog,
-    retry,
-    retrylog,
-    verboseLog
-} from "../../api/utils";
-import { getColor } from "./command-utils";
+import { checkCertificate, getHomeDotMcDir, retry } from "../../api/utils";
+import { errorLog, getColor, homeDirLog, proxyLog, retrylog, verboseLog } from "./command-utils";
 
 const color = getColor("cyan");
