+++ src/cli/commands/mc-create-event.ts
@@ -4,16 +4,8 @@ import * as fs from "fs";
 import * as os from "os";
 import * as path from "path";
 import { MindConnectAgent, MindsphereStandardEvent, retry } from "../..";
-import {
-    checkCertificate,
-    errorLog,
-    getHomeDotMcDir,
-    homeDirLog,
-    proxyLog,
-    retrylog,
-    verboseLog
-} from "../../api/utils";
-import { getColor } from "./command-utils";
+import { checkCertificate, getHomeDotMcDir } from "../../api/utils";
+import { errorLog, getColor, homeDirLog, proxyLog, retrylog, verboseLog } from "./command-utils";
 
 const color = getColor("cyan");
