+++ test/agent-management.spec.ts
@@ -1,8 +1,8 @@
 import * as chai from "chai";
 import { IMindConnectConfiguration, MindConnectAgent } from "../src";
 import { AgentManagementModels, AssetManagementModels, MindSphereSdk } from "../src/api/sdk";
-import { decrypt, loadAuth, throwError } from "../src/api/utils";
 import { sleep } from "./test-utils";
+import { decrypt, throwError, loadAuth } from "../src/api/utils";
 chai.should();
 
 describe("[SDK] AgentManagementClient", () => {
